nouvelle integration page structure
ajouter du changer "autres types"

ajout de la case Label France Services

Update aidants_connect_habilitation/forms.py


init modif suite retours Christophe

nouvelle ecriture

Fix autocomplete

reverse redirection formulaire à la validation

fix test
