+++ aidants_connect_common/templates/layouts/autocomplete.html
@@ -1,7 +1,5 @@
 {% load static %}
 {% if not GOUV_ADDRESS_SEARCH_API_DISABLED %}
   <link rel="stylesheet" href="{% static 'css/autocomplete.css' %}">
-  <script defer src="{{ AUTOCOMPLETE_SCRIPT_SRC }}"></script>
-  <script defer src="{% static 'js/abstract_autocomplete.js' %}"></script>
-  <script defer src="{% static 'js/address_autocomplete.js' %}"></script>
+  <script type="module" src="{% static 'js/address-autocomplete.mjs' %}"></script>
 {% endif %}
