+++ aidants_connect_habilitation/static/js/organisation-form.mjs
@@ -4,26 +4,19 @@ class DynamicForm extends BaseController {
     static targets = [
         "typeOtherInputContainer",
         "typeInput",
-        "onlyShownIfPrivateOrg",
-        "privateOrgInput",
-        "franceServicesInput",
-        "onlyShownIfFranceServices"
+        "onlyShownIfFranceServices",
+        "franceServicesInput"
     ];
 
     static values = {typeOther: String}
 
     connect() {
         this.mutateVisibility(this.typeInputTarget.value === this.typeOtherValue, this.typeOtherInputContainerTarget);
-        this.mutateVisibility(this.privateOrgInputTarget.checked, this.onlyShownIfPrivateOrgTarget);
         this.mutateVisibility(this.franceServicesInputTarget.checked, this.onlyShownIfFranceServicesTarget);
     }
 
      onFranceServicesChange(evt) {
         this.mutateVisibility(evt.target.checked, this.onlyShownIfFranceServicesTarget);
-    }
-
-    onIsPrivateOrgChange(evt) {
-        this.mutateVisibility(evt.target.checked, this.onlyShownIfPrivateOrgTarget);
     }
 
     onTypeChange(evt) {
