+++ aidants_connect_habilitation/templates/aidants_connect_habilitation/forms/organisation.html
@@ -1,0 +1,89 @@
+{% load ac_common dsfr_tags widget_tweaks %}
+{% if errors %}
+  <section class="fr-my-4v fr-input-group fr-input-group--error">
+    {{ errors }}
+  </section>
+{% endif %}
+<fieldset class="fr-fieldset">
+  <legend class="fr-fieldset__legend">Informations générales</legend>
+  <div class="fr-fieldset__element">
+    <p>Vous pouvez faire autant de demandes qu’il existe de structures d’accueil.</p>
+  </div>
+  <div class="fr-fieldset__element">
+    {% dsfr_form_field form.type|attr:"data-action:change->dynamic-form#onTypeChange"|attr:"data-dynamic-form-target:typeInput" %}
+  </div>
+  <div class="fr-fieldset__element" data-dynamic-form-target="typeOtherInputContainer">
+    {% dsfr_form_field form.type_other %}
+  </div>
+  <div class="fr-fieldset__element">
+    {% dsfr_form_field form.france_services_label|attr:"data-action:change->dynamic-form#onFranceServicesChange"|attr:"data-dynamic-form-target:franceServicesInput" %}</div>
+  <div
+    class="fr-fieldset__element"
+    data-dynamic-form-target="onlyShownIfFranceServices"
+  >
+    {% dsfr_form_field form.france_services_number %}
+  </div>
+  <div class="fr-fieldset__element">{% dsfr_form_field form.name %}</div>
+  <div class="fr-fieldset__element">{% dsfr_form_field form.siret %}</div>
+  <p>
+    <em>Besoin d’aide pour retrouver votre SIRET ?</em>
+    <a
+      class="fr-link fr-link--icon-right fr-icon-arrow-right-line"
+      href="https://annuaire-entreprises.data.gouv.fr/?mtm_campaign=aidants_connect"
+    >
+      Annuaire des Entreprises
+    </a>
+  </p>
+</fieldset>
+<fieldset class="fr-fieldset">
+  <legend class="fr-fieldset__legend">Informations descriptives</legend>
+  <div class="fr-fieldset__element">{% dsfr_form_field form.web_site %}</div>
+  <div class="fr-fieldset__element">{% dsfr_form_field form.mission_description %}</div>
+  <div class="fr-fieldset__element">{% dsfr_form_field form.avg_nb_demarches %}</div>
+
+</fieldset>
+<fieldset
+  class="fr-fieldset"
+  {% if not GOUV_ADDRESS_SEARCH_API_DISABLED %}
+  data-controller="address-autocomplete"
+  data-address-autocomplete-api-base-url-value="{{ GOUV_ADDRESS_SEARCH_API_BASE_URL }}"
+  {% endif %}
+>
+  <legend class="fr-fieldset__legend">Informations administratives</legend>
+  <div class="fr-fieldset__element">
+    <p>L’adresse saisie ci-dessous sera celle utilisée sur les mandats édités par Aidants Connect.</p>
+  </div>
+  {% if form.should_display_addresses_select %}
+    <div class="fr-fieldset__element">
+      {% dsfr_form_field form.alternative_address %}
+    </div>
+  {% endif %}
+  <div class="fr-fieldset__element">{% dsfr_form_field form.address %}</div>
+  <div class="fr-fieldset__element fr-fieldset__element--inline fr-fieldset__element--inline-grow">
+    {% dsfr_form_field form.zipcode %}
+  </div>
+  <div class="fr-fieldset__element fr-fieldset__element--inline fr-fieldset__element--inline-grow">
+    {% dsfr_form_field form.city %}
+  </div>
+  {% with errors=form.zipcode.errors|add:form.city.errors %}
+    <div class="fr-grid-row fr-grid-row--gutters zipcode-container">
+      <div class="fr-col-8 fr-col-md-5">
+        {{ form.city_insee_code }}
+        {{ form.department_insee_code }}
+      </div>
+    </div>
+    {% if errors %}
+      <div class="fr-grid-row fr-grid-row--gutters">
+        <div class="errors fr-col-12 fr-col-md-7 fr-col-offset-md-5">{{ errors }}</div>
+      </div>
+    {% endif %}
+  {% endwith %}
+  <div role="alert" class="fr-alert fr-alert--info fr-alert--sm fr-mt-3w">
+    <p>
+      Si votre adresse n’est pas reconnue,
+      <a class="fr-link" href="{% mailto_href 'contact@aidantsconnect.beta.gouv.fr' %}">
+        contactez l’équipe Aidants Connect.
+      </a>
+    </p>
+  </div>
+</fieldset>
