+++ aidants_connect_habilitation/templates/aidants_connect_habilitation/validation_request_form_view.html
@@ -103,8 +103,6 @@
             </div>
             <div class="fr-col-12 fr-col-md-6">
               <div class="fr-grid-row">
-                <div class="fr-col-12 fr-col-md-6 fr-mb-2w fr-text-mention--grey">Structure privée</div>
-                <div class="fr-col-12 fr-col-md-6 fr-mb-2w">{{ organisation.is_private_org|yesno:"Oui,Non" }}</div>
                 <div class="fr-col-12 fr-col-md-6 fr-mb-2w fr-text-mention--grey">Label France Services</div>
                 <div class="fr-col-12 fr-col-md-6 fr-mb-2w">{{ organisation.france_services_label|yesno:"Oui,Non" }}</div>
               </div>
