+++ aidants_connect_habilitation/templates/organisation_form.html
@@ -1,106 +1,42 @@
-{% extends 'layouts/main-habilitation.html' %}
-{% load ac_common form_extras static %}
+{% extends 'layouts/main.html' %}
+{% load ac_common dsfr_tags form_extras static %}
 
 {% block title %}
-  Aidants Connect - Structure
+  Structure - Aidants Connect
 {% endblock title %}
 
 {% block content %}
   <div class="fr-container">
-    {% include 'edito/_titre_formulaire_habilitation.html' %}
-    {% include "layouts/_breadcrumbs-legacy.html" %}
+    <div class="fr-mb-6w">
+      <h1 class="fr-mb-2v">Formulaire d’habilitation</h1>
+    </div>
     <div class="fr-grid-row fr-grid-row--gutters">
-
-      <div class="fr-col-12 fr-col-md-8">
+      <div class="fr-col-12 fr-col-md-6">
+        {% include "layouts/_breadcrumbs.html" %}
         <form
           method="post"
+          class="fr-my-8v"
           data-controller="dynamic-form"
           data-dynamic-form-type-other-value="{{ type_other_value }}"
         >
-          <h2>Structure d’accueil</h2>
-          <p class="subtitle">Vous pouvez faire autant de demandes qu’il existe de structures d’accueil. Tous les champs
-            sont obligatoires.</p>
-          <div class="shadowed padding-2rem">
-            <h3>Informations générales</h3>
-            {% csrf_token %}
-            {{ form.non_field_errors }}
-            {% field_as_fr_grid_row form.type %}
-            <div data-dynamic-form-target="typeOtherInputContainer">
-              {% field_as_fr_grid_row form.type_other %}
-            </div>
-            {% field_as_fr_grid_row form.name %}
-            <h3>Informations administratives</h3>
-
-            <section
-              data-controller="address-autocomplete"
-              data-address-autocomplete-api-base-url-value="{{ GOUV_ADDRESS_SEARCH_API_BASE_URL }}"
-            >
-              {% include "_address_validatable_mixin.html" with form=form %}
-
-              {% field_as_fr_grid_row form.address %}
-
-              {% with errors=form.zipcode.errors|add:form.city.errors %}
-                <div
-                  class="form-grid-row fr-grid-row fr-grid-row--gutters {% if errors %}form-grid-row-error{% endif %}">
-                  <div class="fr-col-12 fr-col-md-5">
-                    <label for="{{ form.zipcode.id_for_label }}">{{ form.zipcode.label }}</label>&nbsp;/&nbsp;
-                    <label for="{{ form.city.id_for_label }}">{{ form.city.label }}</label>
-                  </div>
-                  <div class="fr-col-4 fr-col-md-2 zipcode-container">
-                    {{ form.zipcode }}
-                  </div>
-                  <div class="fr-col-8 fr-col-md-5">
-                    {{ form.city }}
-                  </div>
-                  <div class="fr-col-8 fr-col-md-5">
-                    {{ form.city_insee_code }}
-                    {{ form.department_insee_code }}
-                  </div>
-                </div>
-
-                {% if errors %}
-                  <div class="fr-grid-row fr-grid-row--gutters">
-                    <div class="errors fr-col-12 fr-col-md-7 fr-col-offset-md-5">{{ errors }}</div>
-                  </div>
-                {% endif %}
-              {% endwith %}
-
-              {% include "layouts/autocomplete.html" %}
-            </section>
-
-            {% field_as_fr_grid_row form.siret %}
-            <p><i>Besoin d’aide pour retrouver votre SIRET ? </i> <a href="https://annuaire-entreprises.data.gouv.fr/?mtm_campaign=aidants_connect">Annuaire des Entreprises →</a></p>
-
-            {% checkbox_fr_grid_row form.is_private_org %}
-
-            <section class="grey-background shown-on-checkbox" data-dynamic-form-target="onlyShownIfPrivateOrg">
-              {% field_as_fr_grid_row form.partner_administration %}
-            </section>
-
-            <h3>Informations descriptives</h3>
-
-            {% field_as_fr_grid_row form.web_site %}
-            {% field_as_fr_grid_row form.mission_description %}
-            {% field_as_fr_grid_row form.avg_nb_demarches %}
-
-            <h3>Autres caractéristiques de la structure</h3>
-
-            {% checkbox_fr_grid_row form.france_services_label %}
-
-            <section class="grey-background shown-on-checkbox" data-dynamic-form-target="onlyShownIfFranceServices">
-              {% field_as_fr_grid_row form.france_services_number %}
-            </section>
-            <div class="button-box">
-              <a class="fr-btn fr-btn--secondary" href="{% url 'habilitation_modify_issuer' issuer_id=issuer.issuer_id %}">
+          {% csrf_token %}
+          {{ form }}
+          <ul class="fr-btns-group fr-btns-group--inline-md fr-my-6w">
+            <li>
+              <a class="fr-btn fr-btn--secondary"
+                href="">
                 Revenir à l’étape précédente
               </a>
-              <button type="submit" class="fr-btn">Valider cette étape</button>
-            </div>
-          </div>
+            </li>
+            <li>
+              <input type="submit" class="fr-btn" value="Passer à l’étape suivante">
+            </li>
+          </ul>
         </form>
+
+        {% dsfr_alert type="info" extra_classes="fr-alert--sm" content="Les informations saisies sont sauvegardées à chaque étape. Vous pouvez revenir plus tard en cliquant sur le lien reçu par email." %}
       </div>
     </div>
-    {% include "_more-info.html" %}
   </div>
 {% endblock content %}
 
@@ -108,7 +44,3 @@
   {% stimulusjs %}
   <script defer type="module" src="{% static 'js/organisation-form.mjs' %}"></script>
 {% endblock extrajs %}
-
-{% block extracss %}
-  <link href="{% static 'css/personnel_form.css' %}" rel="stylesheet">
-{% endblock extracss %}
